HT_19_django_celery_proj/requirements.txt

dj-database-url==0.5.0
Django==3.1.3
django-bootstrap4==2.3.1
requests==2.24.0
whitenoise==5.2.0
celery==5.0.5
django-db==0.0.7
pymysql==1.0.7
python-crontab==2.5.1
django_celery_results==2.0.1
django-celery-beat==2.2.0


